pyproject.toml
[project]
name = "news-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "feedparser>=6.0.11",
    "jinja2>=3.1.6",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "praw>=7.7.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "google-genai>=1.51.0",
    "markdown>=3.8",
    "pydantic-settings>=2.9.1",
    "tenacity>=9.1.2",
    "markupsafe>=3.0.2",
    "crawl4ai>=0.7.4,<0.8",
    "alembic>=1.16.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.75.0",
    "ffmpeg-python>=0.2.0",
    "python-dateutil>=2.9.0.post0",
    "playwright>=1.53.0",
    "yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git",
    "bgutil-ytdlp-pot-provider>=1.2.2",
    "openai-whisper>=20250625",
    "jmespath>=1.0.1",
    "torch>=2.8.0",
    "pyjwt[crypto]",
    "passlib[bcrypt]",
    "authlib>=1.6.5",
    "email-validator>=2.3.0",
    "anthropic>=0.72.0",
    "rich>=14.2.0",
    "pydantic-ai>=1.2.1",
    "exa-py>=1.0.0",
    "elevenlabs>=2.35.0",
    "pillow>=12.0.0",
    "trafilatura>=2.0.0",
    "langfuse>=3.14.3",
]

[project.scripts]
newsly-agent = "cli.newsly_agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "cli*", "cron*"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "node_modules",
    "tests",
]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["B008"]  # Depends() in function arguments is standard FastAPI pattern

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*" = ["E501"]
"migrations/*" = ["E501"]
"app/services/openai_llm.py" = ["E501", "SIM116"]
"app/services/google_flash.py" = ["E501", "SIM116"]
